inzure_back/requirements.txt
Django==1.10.4
django-appconf==1.0.2
django-compressor==2.1
django-material==0.11.0
djangorestframework==3.5.3
drf-nested-routers==0.11.1
rcssmin==1.0.6
rjsmin==1.0.12
